+++ packages/react/src/components/cookieConsent/types.ts
@@ -1,0 +1,3 @@
+import { CookieContentSource as ImportedContentSource } from './content.builder';
+
+export type ContentSource = ImportedContentSource;